Move handleFileContent to WmskmlController
