+++ contribs/gmf/apps/desktop/js/controller.js
@@ -26,15 +26,13 @@ goog.require('app.GmfImportHelper');
 /**
  * @param {angular.Scope} $scope Scope.
  * @param {angular.$injector} $injector Main injector.
- * @param {ngeo.File} ngeoFile The file service.
- * @param {gettext} gettext The gettext service
  * @param {angular.$q} $q Angular $q.
  * @constructor
  * @extends {gmf.AbstractDesktopController}
  * @ngInject
  * @export
  */
-app.DesktopController = function($scope, $injector, ngeoFile, gettext, $q) {
+app.DesktopController = function($scope, $injector, $q) {
   gmf.AbstractDesktopController.call(this, {
     srid: 21781,
     mapViewConfig: {
@@ -105,7 +103,7 @@ app.DesktopController = function($scope, $injector, ngeoFile, gettext, $q) {
   /**
    * @export
    */
-  this.importOptions = new app.GmfImportHelper(this.map, $scope, gettext, ngeoFile, $q).createOptions();
+  this.importOptions = new app.GmfImportHelper($q).createOptions();
 };
 ol.inherits(app.DesktopController, gmf.AbstractDesktopController);
