+++ contribs/gmf/apps/desktop/js/importhelper.js
@@ -1,7 +1,5 @@
 goog.provide('app.GmfImportHelper');
 
-goog.require('ol.format.GPX');
-goog.require('ol.format.KML');
 goog.require('ol.source.ImageWMS');
 goog.require('ol.layer.Image');
 goog.require('ol.source.Vector');
@@ -10,21 +8,10 @@ goog.require('ol.layer.Vector');
 
 /**
  * @constructor
- * @param {ol.Map} map The map.
- * @param {angular.Scope} $scope The scope.
- * @param {gettext} gettext .
- * @param {ngeo.File} ngeoFile Ngeo file.
  * @param {angular.$q} $q Q.
  */
-app.GmfImportHelper = function(map, $scope, gettext, ngeoFile, $q) {
-  this.map_ = map;
-  this.$scope_ = $scope;
-  $scope['map'] = map;
-  this.ngeoFile_ = ngeoFile;
+app.GmfImportHelper = function($q) {
   this.$q_ = $q;
-  this.gpxFormat_ = new ol.format.GPX();
-  this.kmlFormat_ = new ol.format.KML();
-  this.gettext_ = gettext;
 
   this.urls = [
     {'name': 'geoadmin', 'url': 'https://wms.geo.admin.ch/'},
@@ -122,84 +109,6 @@ app.GmfImportHelper.prototype.getOlLayerFromGetCapLayer = function(getCapLayer) 
 
 
 /**
- * @param {ol.Map} map The map
- * @param {Array<ol.Feature>} features The features
- */
-app.GmfImportHelper.prototype.addFeatures = function(map, features) {
-  const source = new ol.source.Vector({
-    features
-  });
-  const layer = new ol.layer.Vector({
-    source
-  });
-  map.addLayer(layer);
-  const size = map.getSize();
-  if (size) {
-    map.getView().fit(source.getExtent(), {size, padding: [30, 30, 30, 30]});
-  }
-};
-
-
-/** Manage data depending on the content
- * @param {string} data Content of the file.
- * @param {File} file Informations of the file (if available).
- * @return {angular.$q.Promise} The promise
- */
-app.GmfImportHelper.prototype.handleFileContent = function(data, file) {
-  const map = this.map_;
-  const defer = this.$q_.defer();
-  const $scope = this.$scope_;
-  const ngeoFile = this.ngeoFile_;
-  let features;
-
-  $scope['gpxContent'] = null;
-  $scope['kmlContent'] = null;
-  $scope['wmsGetCap'] = null;
-  $scope['wmtsGetCap'] = null;
-
-  if (ngeoFile.isWmsGetCap(data)) {
-    $scope['wmsGetCap'] = data;
-    defer.resolve({
-      'message': this.gettext_('Download succeeded')
-    });
-  } else if (ngeoFile.isWmtsGetCap(data)) {
-    $scope['wmtsGetCap'] = data;
-    defer.resolve({
-      'message': this.gettext_('Download succeeded')
-    });
-  } else if (ngeoFile.isKml(data)) {
-    features = this.kmlFormat_.readFeatures(data, {
-      featureProjection: map.getView().getProjection()
-    });
-
-    this.addFeatures(map, features);
-    defer.resolve({
-      'message': this.gettext_('Parsing succeeded')
-    });
-  } else if (ngeoFile.isGpx(data)) {
-    features = this.gpxFormat_.readFeatures(data, {
-      featureProjection: map.getView().getProjection()
-    });
-
-    this.addFeatures(map, features);
-    defer.resolve({
-      'message': this.gettext_('Parsing succeeded')
-    });
-
-  } else {
-    console.error('Unparseable content: ', data);
-    defer.reject({
-      'message': this.gettext_('Parsing failed'),
-      'reason': this.gettext_('unsupported format')
-    });
-  }
-  // WMTS
-
-  return defer.promise;
-};
-
-
-/**
  * @return {Object} the options
  */
 app.GmfImportHelper.prototype.createOptions = function() {
@@ -221,7 +130,6 @@ app.GmfImportHelper.prototype.createOptions = function() {
         url += '?SERVICE=WMS&REQUEST=GetCapabilities&VERSION=1.3.0';
       }
       return this.$q_.when(url);
-    },
-    'handleFileContent': this.handleFileContent.bind(this)
+    }
   };
 };
